fix(ADFA-2024): Preserve decimal precision

+++ app/src/main/java/com/itsaky/androidide/utils/StorageSizeUtils.kt
@@ -1,9 +1,9 @@
 package com.itsaky.androidide.utils
 
-private const val BYTES_IN_GIGABYTE = 1024L * 1024L * 1024L
+private const val BYTES_IN_GIGABYTE = 1024F * 1024F * 1024F
 
 /** Converts a Long representing bytes into Gigabytes. */
-fun Long.bytesToGigabytes(): Long = this / BYTES_IN_GIGABYTE
+fun Float.bytesToGigabytes(): Float = this / BYTES_IN_GIGABYTE
 
 /** Converts a Long representing Gigabytes into bytes. */
-fun Long.gigabytesToBytes(): Long = this * BYTES_IN_GIGABYTE
+fun Long.gigabytesToBytes(): Float = this * BYTES_IN_GIGABYTE
